// 06-admin-shop/admin-shop/src/modules/products/interfaces/product.interface.ts
import type { User } from "@/modules/auth/interfaces/User.interface";

export interface Product {
  id:          string;
  title:       string;
  price:       number;
  description: string;
  slug:        string;
  stock:       number;
  sizes:       Size[];
  gender:      string;
  tags:        string[];
  images:      string[];
  user:        User;
}

export enum Size {
  L = "L",
  M = "M",
  S = "S",
  Xl = "XL",
  Xs = "XS",
  Xxl = "XXL",
}